% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_build.R
\name{pkg_build}
\alias{pkg_build}
\title{Build Package}
\usage{
pkg_build(
  pkg_dir = getwd(),
  style = TRUE,
  data = TRUE,
  render = TRUE,
  readme = TRUE,
  vignettes = TRUE,
  tests = TRUE,
  pkgdown = TRUE,
  par = TRUE,
  ncores = NULL
)
}
\arguments{
\item{pkg_dir}{Directory where the package is initialized.}

\item{style}{Logical.
Style R scripts and R Markdown files.}

\item{data}{Logical.
Generate data from \code{data_raw}.}

\item{render}{Logical.
Render R scripts and R Markdown files.}

\item{readme}{Logical.
Render \code{README.Rmd}.
Ignored if \code{render = FALSE}.}

\item{vignettes}{Logical.
Render \code{vignettes}.
Ignored if \code{render = FALSE}.}

\item{tests}{Logical.
Render \code{tests}.
Ignored if \code{render = FALSE}.}

\item{pkgdown}{Logical.
Build \code{pkgdown} site.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.}

\item{ncores}{Integer.
Number of cores to use if \code{par = TRUE}.
If unspecified, defaults to \code{detectCores() - 1}.}
}
\description{
Build Package
}
\author{
Ivan Jacob Agaloos Pesigan
}
