% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_build_minimal.R
\name{pkg_build_minimal}
\alias{pkg_build_minimal}
\title{Build Package (Minimal)}
\usage{
pkg_build_minimal(pkg_root = getwd())
}
\arguments{
\item{pkg_root}{Character string.
Package root directory.}
}
\description{
Minimal check, build, and install
using \verb{R CMD}.
}
\author{
Ivan Jacob Agaloos Pesigan
}
