% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_checkroot.R
\name{pkg_checkroot}
\alias{pkg_checkroot}
\title{Check Package Root}
\usage{
pkg_checkroot(dir = getwd())
}
\arguments{
\item{dir}{Character string.
Directory.}
}
\description{
Check Package Root
}
\examples{
\dontrun{
pkg_checkroot(
  dir = getwd()
)
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
