% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_create.R
\name{pkg_create}
\alias{pkg_create}
\title{Create Boilerplate Package R Package.}
\usage{
pkg_create(
  pkg_dir = getwd(),
  pkg_name = "boilerplatePackage",
  input_file = NULL,
  add_description = NULL,
  add_namespace = NULL,
  add_rbuildignore = NULL,
  add_gitignore = NULL,
  add_travis = NULL,
  git = FALSE,
  github = FALSE
)
}
\arguments{
\item{pkg_dir}{Directory where the package is initialized.}

\item{pkg_name}{Character string.
Package name.}

\item{input_file}{csv file containing \code{DESCRIPTION} fields and entries.}

\item{add_description}{Character string.
Additional entries to the \code{DESCRIPTION} file
not included in \code{input_file}}

\item{add_namespace}{Character string.
Entries to the \code{NAMESPACE}
in addition to the boilerplate example.}

\item{add_rbuildignore}{Character string.
Entries to the \code{.Rbuildignore}
in addition to the boilerplate example.}

\item{add_gitignore}{Character string.
Entries to the \code{.gitignore}
in addition to the boilerplate example.}

\item{add_travis}{Character string.
Entries to the \code{.travis.yml}
in addition to the boilerplate example.}

\item{git}{Logical.
Set up a git repository.}

\item{github}{Logical.
Set up and push to a github repository.}
}
\description{
Create Boilerplate Package R Package.
}
\author{
Ivan Jacob Agaloos Pesigan
}
