% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_data_raw.R
\name{pkg_data_raw}
\alias{pkg_data_raw}
\title{Source \code{R} Files in data_raw.}
\usage{
pkg_data_raw(pkg_root = getwd())
}
\arguments{
\item{pkg_root}{Character string.
Package root directory.}
}
\description{
Sources \code{R} files in data_raw.
}
\examples{
\dontrun{
pkg_data_raw(
  pkg_root = "~/boilerplatePackage"
)
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
