% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_description.R
\name{pkg_description}
\alias{pkg_description}
\title{Create a Package \code{DESCRIPTION} File.}
\usage{
pkg_description(
  pkg_dir = getwd(),
  input_file = NULL,
  add = NULL,
  msg = "DESCRIPTION file path:"
)
}
\arguments{
\item{pkg_dir}{Character string.
Directory where the package is initialized.}

\item{input_file}{Character string.
\code{YAML} file containing \code{DESCRIPTION} fields and entries.}

\item{add}{Character string.
Additional entries to the \code{DESCRIPTION} file
not included in \code{input_file}}

\item{msg}{Character string.
Prints optional message and output directory.}
}
\description{
Creates a package \code{DESCRIPTION} file
by extracting information
from an external \code{yml} \code{input_file}.
See \code{system.file("extdata", "DESCRIPTION.yml", package = "jeksterslabRpkg", mustWork = TRUE)}
for the \code{input_file} template.
}
\details{
Note that if \code{\link[jeksterslabRpkg:pkg_create]{jeksterslabRpkg::pkg_create()}} is used,
this function will be called.

\strong{THIS FUNCTION OVERWRITES AN EXISTING \code{DESCRIPTION} FILE
IN THE SPECIFIED PACKAGE ROOT DIRECTORY.
USE WITH CAUTION.}
}
\examples{
\dontrun{
pkg_description(
  pkg_dir = getwd(),
  input_file = "DESCRIPTION.yml"
)
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
