% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_description_yml.R
\name{pkg_description_yml}
\alias{pkg_description_yml}
\title{Parse Description \code{YAML}}
\usage{
pkg_description_yml(
  input_file = NULL,
  fields = c("Package", "Given", "Family", "Email", "ORCID", "Version", "Title",
    "Description", "License", "Github", "Encoding", "Language", "LazyData",
    "VignetteBuilder", "Roxygen"),
  required = TRUE,
  dependencies = TRUE
)
}
\arguments{
\item{input_file}{Character string.
\code{YAML} file containing \code{DESCRIPTION} fields and entries.}

\item{fields}{Character vector.
Fields from \code{DESCRIPTION.yml} to extract.
Note that this only works for fields that contain a single value.}

\item{required}{Logical.
Check required fields.}

\item{dependencies}{Logical.
Extract dependencies
(\code{SystemRequirements}, \code{Depends}, \code{Imports}, \code{Suggests}).}
}
\description{
Parse Description \code{YAML}
}
\examples{
\dontrun{
pkg_description_yml(
  input_file = "DESCRIPTION.yml",
  fields = c(
    "Package",
    "Given",
    "Family",
    "Email",
    "ORCID",
    "Version",
    "Title",
    "Description",
    "License",
    "Github",
    "Encoding",
    "Language",
    "LazyData",
    "VignetteBuilder",
    "Roxygen"
  ),
  required = TRUE,
  dependencies = TRUE
)
}
}
