% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_find_root.R
\name{pkg_find_root}
\alias{pkg_find_root}
\title{Find Package Root Directory}
\usage{
pkg_find_root(dir = getwd(), pkg_name, par = TRUE, ncores = NULL)
}
\arguments{
\item{dir}{Character string.
Directory to search.
If unspecified, defaults to \code{Sys.getenv("HOME")}.}

\item{pkg_name}{Character string.
Package name.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.}

\item{ncores}{Integer.
Number of cores to use if \code{par = TRUE}.
If unspecified, defaults to \code{detectCores() - 1}.}
}
\description{
Finds package root directory by seraching for
tne DESCRIPTION file that contains
\code{Package: pkg_name}.
}
\details{
The function starts searching the specified \code{dir}.
If the search does not return the root directory
for the specified \code{pkg_name},
If the function fails to find a valid package root directory in dir,
the following directories will be searched:
\itemize{
\item \code{../dir}
\item \code{../../dir}
\item \verb{working directory} (\code{getwd()})
\item \code{Sys.getenv("HOME")}
\item \code{/}
}
}
\examples{
\dontrun{
pkg_find_root(
  dir = getwd(),
  pkg_name = "boilerplatePackage",
  par = FALSE
)
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
