% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_getstarted.R
\name{pkg_getstarted}
\alias{pkg_getstarted}
\title{Create a Boilerplate Package Get Started Vignette.}
\usage{
pkg_getstarted(pkg_root, msg = "Get started file path:")
}
\arguments{
\item{pkg_root}{Character string.
Package root directory.}
}
\description{
Creates a boilerplate package Get Started vignette.
}
\details{
Note that if \code{\link[=pkg_create]{pkg_create()}} is used,
this function will be called.
}
\examples{
\dontrun{
pkg_getstarted(
  pkg_root = "~/boilerplatePackage"
)
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
