% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_git.R
\name{pkg_git}
\alias{pkg_git}
\title{Git and Github}
\usage{
pkg_git(pkg_root, github = TRUE, msg = "BUILD")
}
\arguments{
\item{pkg_root}{Character string.
Package root directory.}

\item{github}{Logical.
Set up and push to a github repository.}

\item{msg}{Character string. Commit message.}
}
\description{
Set up a \code{Git} repository
and push the created repo to \code{Github}.
This requires that \code{git} and \code{hub} are installed
and configured in the system.
}
\author{
Ivan Jacob Agaloos Pesigan
}
