% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_gitignore.R
\name{pkg_gitignore}
\alias{pkg_gitignore}
\title{Create a Boilerplate Package \code{.gitignore} File.}
\usage{
pkg_gitignore(pkg_root, add = NULL, msg = ".gitignore file path:")
}
\arguments{
\item{pkg_root}{Character string.
Package root directory.}

\item{add}{Character string.
Entries to the \code{.gitignore}
in addition to the boilerplate example.}

\item{msg}{Character string.
Optional message.
If supplied,
prints \verb{msg dir/fn}.}
}
\description{
Creates a boilerplate package \code{.gitignore} file.
}
\details{
Note that if \code{\link[jeksterslabRpkg:pkg_create]{jeksterslabRpkg::pkg_create()}} is used,
this function will be called.

If \code{.gitignore} exists in the specified package root directory,
it will \strong{NOT} be overwritten.
Additional entries in the \code{add} argument will be appended.
}
\examples{
\dontrun{
pkg_gitignore(
  pkg_root = getwd()
)
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
