% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_gitlsf.R
\name{pkg_gitlsf}
\alias{pkg_gitlsf}
\title{Create a Boilerplate Package \code{.gitattributes} File with Git LFS Settings.}
\usage{
pkg_gitlsf(
  pkg_dir = getwd(),
  pkg_name,
  ext = c("gz", "zip", "Rds", "rds", "Rda", "rda", "jpg", "jpeg", "png", "pdf")
)
}
\arguments{
\item{pkg_dir}{Character string.
Directory where the package is initialized.}

\item{pkg_name}{Character string.
Package name.}

\item{ext}{Character vector.
File extensions.}
}
\description{
Create a Boilerplate Package \code{.gitattributes} File with Git LFS Settings.
}
\examples{
\dontrun{
pkg_gitlfs(
  pkg_root = getwd(),
  ext = c("pdf", "jpg")
)
}
}
