% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_license.R
\name{pkg_license}
\alias{pkg_license}
\title{Create a Package License.}
\usage{
pkg_license(pkg_root, input_file = NULL, msg = "LICENSE file path:")
}
\arguments{
\item{pkg_root}{Character string.
Package root directory.}

\item{input_file}{Character string.
\code{YAML} file containing \code{DESCRIPTION} fields and entries.}

\item{msg}{Character string.
Prints optional message and output directory.}
}
\description{
Creates a package \code{LICENSE} and \code{LICENSE.md}
by extracting information
from an external \code{yml} \code{input_file}.
See \code{system.file("extdata", "DESCRIPTION.yml", package = "jeksterslabRpkg", mustWork = TRUE)}
for the \code{input_file} template.
}
\details{
Note that if \code{\link[jeksterslabRpkg:pkg_create]{jeksterslabRpkg::pkg_create()}} is used,
this function will be called.

\strong{THIS FUNCTION OVERWRITES EXISTING \code{LICENSE} and \code{LICENSE.md} FILES
IN THE SPECIFIED PACKAGE ROOT DIRECTORY.
USE WITH CAUTION.}
}
\examples{
\dontrun{
pkg_license(
  pkg_root = getwd(),
  input_file = "DESCRIPTION.yml"
)
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
