% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_namespace.R
\name{pkg_namespace}
\alias{pkg_namespace}
\title{Create a Boilerplate Package \code{NAMESPACE}.}
\usage{
pkg_namespace(pkg_dir = getwd(), pkg_name, add = NULL)
}
\arguments{
\item{pkg_dir}{Character string.
Directory where the package is initialized.}

\item{pkg_name}{Character string.
Package name.}

\item{add}{Character string.
Entries to the \code{NAMESPACE}
in addition to the boilerplate example.}
}
\description{
Note that if \code{\link[jeksterslabRpkg:pkg_create]{jeksterslabRpkg::pkg_create()}} is used,
this function will be called.
}
\details{
If you are going to document, check, and build your package using \code{devtools},
you may delete the boillerplate \code{NAMESPACE} file.

\strong{THIS FUNCTION OVERWRITES AN EXISTING \code{NAMESPACE} FILE
IN THE SPECIFIED PACKAGE ROOT DIRECTORY.
USE WITH CAUTION.}
}
\examples{
\dontrun{
pkg_namespace(
  pkg_dir = getwd(),
  pkg_name = "boilerplatePackage"
)
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
