% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_news.R
\name{pkg_news}
\alias{pkg_news}
\title{Create a Boilerplate Package NEWS.md File.}
\usage{
pkg_news(pkg_root, msg = "NEWS.md file path:")
}
\arguments{
\item{pkg_root}{Character string.
Package root directory.}

\item{msg}{Character string.
Optional message.
If supplied,
prints \verb{msg dir/fn}.}
}
\description{
Creates a boilerplate package NEWS.md file.
}
\details{
Note that if \code{\link[jeksterslabRpkg:pkg_create]{jeksterslabRpkg::pkg_create()}} is used,
this function will be called.
}
\examples{
\dontrun{
pkg_news(
  pkg_root = "~/boilerplatePackage"
)
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
