% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_pkgdown.R
\name{pkg_pkgdown}
\alias{pkg_pkgdown}
\title{Create Boilerplate Package .travis.yml.}
\usage{
pkg_pkgdown(pkg_dir = getwd(), pkg_name, input_file = NULL)
}
\arguments{
\item{pkg_dir}{Directory where the package is initialized.}

\item{pkg_name}{Character string.
Package name.}

\item{input_file}{csv file containing \code{DESCRIPTION} fields and entries.}
}
\description{
Creates .travis.yml with \code{codecov}.
}
\examples{
pkg_pkgdown(pkg_dir = tempdir(), pkg_name = "boilerplatePackage")
}
\author{
Ivan Jacob Agaloos Pesigan
}
