% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_rd.R
\name{pkg_rd}
\alias{pkg_rd}
\title{Create a Boilerplate Package \code{Rd} Function Documentation}
\usage{
pkg_rd(pkg_root, msg = "z.Rd file path:")
}
\arguments{
\item{pkg_root}{Character string.
Package root directory.}

\item{msg}{Character string.
Optional message.
If supplied,
prints \verb{msg dir/fn}.}
}
\description{
Creates a boilerplate package \code{Rd} function documentation.
}
\details{
Note that if \code{\link[=pkg_create]{pkg_create()}} is used,
this function will be called.

If you are going to document, check, and build your package using \code{devtools},
you may delete the \code{Rd} function documentation file generated by this function.
}
\examples{
\dontrun{
pkg_rd(
  pkg_root = "~/boilerplatePackage"
)
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
