% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_readme.R
\name{pkg_readme}
\alias{pkg_readme}
\title{Create a Boilerplate Package README File}
\usage{
pkg_readme(pkg_root, input_file = NULL, msg = "README.Rmd file path:")
}
\arguments{
\item{pkg_root}{Character string.
Package root directory.}

\item{input_file}{Character string.
\code{YAML} file containing \code{DESCRIPTION} fields and entries.}

\item{msg}{Character string.
Optional message.
If supplied,
prints \verb{msg dir/fn}.}
}
\description{
Creates a boilerplate package \code{README.Rmd} file
by extracting information
from an external \code{yml} \code{input_file}.
See \code{system.file("extdata", "DESCRIPTION.yml", package = "jeksterslabRpkg", mustWork = TRUE)}
for the \code{input_file} template.
}
\details{
Note that if \code{\link[=pkg_create]{pkg_create()}} is used,
this function will be called.

\strong{THIS FUNCTION OVERWRITES AN EXISTING \code{README.Rmd} FILE
IN THE SPECIFIED PACKAGE ROOT DIRECTORY.
USE WITH CAUTION.}
}
\examples{
\dontrun{
pkg_readme(
  pkg_root = "~/boilerplatePackage",
  input_file = "DESCRIPTION.yml"
)
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
