% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_render.R
\name{pkg_render}
\alias{pkg_render}
\title{Render \code{R} Markdown Files in the Package Directory.}
\usage{
pkg_render(
  pkg_root = getwd(),
  readme = TRUE,
  vignettes = TRUE,
  tests = TRUE,
  par = TRUE,
  ncores = NULL
)
}
\arguments{
\item{pkg_root}{Character string.
Package root directory.}

\item{readme}{Logical.
Render \code{README.Rmd}.}

\item{vignettes}{Logical.
Render \verb{vignettes/*.Rmd}.}

\item{tests}{Logical.
Render \verb{tests/testhat/*.R}.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.}

\item{ncores}{Integer.
Number of cores to use if \code{par = TRUE}.
If unspecified, defaults to \code{detectCores() - 1}.}
}
\description{
Render \code{R} Markdown Files in the Package Directory.
}
\examples{
\dontrun{
pkg_render(
  pkg_root = getwd(),
  readme = TRUE,
  vignettes = TRUE,
  tests = TRUE,
  par = FALSE
)
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
