% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_rmd.R
\name{pkg_rmd}
\alias{pkg_rmd}
\title{Render R Markdown Files in the Package Directory.}
\usage{
pkg_rmd(pkg_dir = getwd())
}
\arguments{
\item{pkg_dir}{Directory where the package is initialized.}
}
\description{
Render R Markdown Files in the Package Directory.
}
\examples{
\dontrun{
pkg_rmd()
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
