% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_style.R
\name{pkg_style}
\alias{pkg_style}
\title{Style Package R and RMD Files.}
\usage{
pkg_style(pkg_dir = getwd())
}
\arguments{
\item{pkg_dir}{Directory where the package is initialized.}
}
\description{
Styles all \code{R} scripts and \code{R Markdown} files
in the package directory.
}
\examples{
\dontrun{
pkg_style()
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
