% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_tests2vignettes.R
\name{pkg_tests2vignettes}
\alias{pkg_tests2vignettes}
\title{Tests to Vignettes}
\usage{
pkg_tests2vignettes(pkg_root = getwd(), par = TRUE, ncores = NULL)
}
\arguments{
\item{pkg_root}{Character string.
Package root directory.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.}

\item{ncores}{Integer.
Number of cores to use if \code{par = TRUE}.
If unspecified, defaults to \code{detectCores() - 1}.}
}
\description{
Copy and spin \verb{tests/testhat/*.R}
to \verb{vignettes/tests/*.Rmd}.
It is assumed that \verb{tests/testhat/*.R}
are written using \code{Roxygen} comments.
}
\author{
Ivan Jacob Agaloos Pesigan
}
