% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_travis.R
\name{pkg_travis}
\alias{pkg_travis}
\title{Create a Boilerplate Package \code{.travis.yml} File.}
\usage{
pkg_travis(pkg_root, add = NULL, msg = ".travis.yml file path:")
}
\arguments{
\item{pkg_root}{Character string.
Package root directory.}

\item{add}{Character string.
Entries to the \code{.travis.yml}
in addition to the boilerplate example.}
}
\description{
Creates a \code{.travis.yml} build file
with additional entries for \code{codecov}.
}
\details{
Note that if \code{\link[=pkg_create]{pkg_create()}} is used,
this function will be called.

\strong{THIS FUNCTION OVERWRITES AN EXISTING \code{.travis.yml} FILE
IN THE SPECIFIED PACKAGE ROOT DIRECTORY.
USE WITH CAUTION.}
}
\examples{
\dontrun{
pkg_travis(
  pkg_root = "~/boilerplatePackage"
)
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
