% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_valid_name.R
\name{pkg_valid_name}
\alias{pkg_valid_name}
\title{Validate Package Name}
\usage{
pkg_valid_name(pkg_name)
}
\arguments{
\item{pkg_name}{Character string.
Package name.}
}
\description{
Validate Package Name
}
\author{
Ivan Jacob Agaloos Pesigan
}
