% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_vignettes.R
\name{pkg_vignettes}
\alias{pkg_vignettes}
\title{Create Boilerplate Package Vignette.}
\usage{
pkg_vignettes(pkg_dir = getwd(), pkg_name)
}
\arguments{
\item{pkg_dir}{Directory where the package is initialized.}

\item{pkg_name}{Character string.
Package name.}
}
\description{
Create Boilerplate Package Vignette.
}
\author{
Ivan Jacob Agaloos Pesigan
}
