% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_aliases.R
\name{rins}
\alias{rins}
\title{Alias for devtools::install()}
\usage{
rins()
}
\description{
Alias for
\code{\link[devtools:install]{devtools::install()}}
with default arguments.
}
\seealso{
\code{\link[devtools:install]{devtools::install()}}
}
