% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_aliases.R
\name{rpkg}
\alias{rpkg}
\title{Aliases jeksterslabRpkg::util_render() Recursive and Parallel}
\usage{
rpkg()

rpkg()
}
\description{
Aliases jeksterslabRpkg::util_render() Recursive and Parallel

Aliases styler::style_pkg(); devtools::check(); devtools::install()
}
