% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_aliases.R
\name{rrender}
\alias{rrender}
\title{Alias for jeksterslabRutils::util_render()}
\usage{
rrender(
  dir = getwd(),
  recursive = TRUE,
  files = NULL,
  par = TRUE,
  ncores = NULL
)
}
\arguments{
\item{dir}{Character string.
Directory which contains target files.
Ignored if \code{file} is NOT NULL.}

\item{recursive}{Logical.
If \code{TRUE},
recursively render\\
all \code{R} scripts (\code{.R}, \code{.r}) and
R Markdown files (\code{.Rmd}, \code{.rmd})
in \code{dir}.
Ignored if \code{file} is NOT NULL.}

\item{files}{Character vector.
Vector of files to render.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.}

\item{ncores}{Integer.
Number of cores to use if
\code{par = TRUE}.
If unspecified,
defaults to
\code{detectCores() - 1}.}
}
\description{
Alias for \code{\link[jeksterslabRutils:util_render]{jeksterslabRutils::util_render()}}
with the following arguments:
\itemize{
\item \code{dir = getwd()}
\item \code{recursive = TRUE}
\item \code{files = NULL}
\item \code{par = TRUE}
\item \code{ncores = NULL}
}
}
\seealso{
\code{\link[jeksterslabRutils:util_render]{jeksterslabRutils::util_render()}}
}
