% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_aliases.R
\name{rsty}
\alias{rsty}
\title{Alias for styler::style_dir()}
\usage{
rsty(
  path = ".",
  ...,
  style = styler::tidyverse_style,
  transformers = style(...),
  filetype = c("R", "Rprofile"),
  recursive = TRUE,
  exclude_files = NULL,
  exclude_dirs = c("packrat", "renv"),
  include_roxygen_examples = TRUE
)
}
\arguments{
\item{path}{Path to a directory with files to transform.}

\item{...}{Arguments passed on to the \code{style} function.}

\item{style}{A function that creates a style guide to use, by default
\code{\link[styler:tidyverse_style]{tidyverse_style()}} (without the parentheses). Not used
further except to construct the argument \code{transformers}. See
\code{\link[styler:style_guides]{style_guides()}} for details.}

\item{transformers}{A set of transformer functions. This argument is most
conveniently constructed via the \code{style} argument and \code{...}. See
'Examples'.}

\item{filetype}{Vector of file extensions indicating which file types should
be styled. Case is ignored, and the \code{.} is optional, e.g.
\code{c(".R", ".Rmd")}, or \code{c("r", "rmd")}. Supported values (after
standardization) are: "r", "rprofile", "rmd", "rnw".}

\item{recursive}{A logical value indicating whether or not files in subdirectories
of \code{path} should be styled as well.}

\item{exclude_files}{Character vector with paths to files that should be
excluded from styling.}

\item{exclude_dirs}{Character vector with directories to exclude.}

\item{include_roxygen_examples}{Whether or not to style code in roxygen
examples.}
}
\description{
Alias for styler::style_dir()
}
\seealso{
\code{\link[styler:style_dir]{styler::style_dir()}}
}
