% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_aliases.R
\name{rtest}
\alias{rtest}
\title{Alias for devtools::test()}
\usage{
rtest(
  pkg = ".",
  filter = NULL,
  stop_on_failure = FALSE,
  export_all = TRUE,
  ...
)
}
\arguments{
\item{pkg}{The package to use, can be a file path to the package or a
package object.  See \code{\link[devtools:as.package]{as.package()}} for more information.}

\item{filter}{If not \code{NULL}, only tests with file names matching this
regular expression will be executed. Matching be performed on the file
name after it has been stripped of \code{"test-"} and \code{".R"}.}

\item{stop_on_failure}{If \code{TRUE}, throw an error if any tests fail.

For historical reasons, the default value of \code{stop_on_failure} is \code{TRUE}
for  \code{test_package()} and \code{test_check()} but \code{FALSE} for \code{test_dir()}, so
if you're calling \code{test_dir()} you may want to consider explicitly setting
\code{stop_on_failure = TRUE}.}

\item{export_all}{If \code{TRUE} (the default), export all objects.
If \code{FALSE}, export only the objects that are listed as exports
in the NAMESPACE file.}

\item{...}{additional arguments passed to \code{\link[testthat:test_dir]{testthat::test_dir()}} and
\code{\link[covr:package_coverage]{covr::package_coverage()}}}
}
\description{
Alias for devtools::test()
}
\seealso{
\code{\link[devtools:test]{devtools::test()}}
}
