% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_appveyor.R
\name{pkg_appveyor}
\alias{pkg_appveyor}
\title{Create a Boilerplate Package \code{appveyor.yml} File.}
\usage{
pkg_appveyor(pkg_root, add = NULL, msg = "appveyor.yml file path:")
}
\arguments{
\item{pkg_root}{Character string.
Package root directory.}

\item{add}{Character string.
Entries to the \code{appveyor.yml}
in addition to the boilerplate example.}

\item{msg}{Character string.
Optional message.
If supplied,
prints \verb{msg dir/fn}.}
}
\description{
Creates an \code{appveyor.yml} build file.
}
\details{
Note that if \code{\link[jeksterslabRpkg:pkg_create]{jeksterslabRpkg::pkg_create()}} is used,
this function will be called.

\strong{THIS FUNCTION OVERWRITES AN EXISTING \code{appveyor.yml} FILE
IN THE SPECIFIED PACKAGE ROOT DIRECTORY.
USE WITH CAUTION.}
}
\examples{
\dontrun{
pkg_appveyor(
  pkg_root = "~/boilerplatePackage"
)
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
