% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_build.R
\name{pkg_build}
\alias{pkg_build}
\title{Build Package.}
\usage{
pkg_build(
  pkg_root = getwd(),
  clean = TRUE,
  minimal = FALSE,
  style = TRUE,
  data = FALSE,
  render = TRUE,
  readme = TRUE,
  vignettes = TRUE,
  tests = TRUE,
  tests2vignettes = TRUE,
  pkgdown = TRUE,
  par = TRUE,
  ncores = NULL,
  git = FALSE,
  github = FALSE,
  commit_msg = "BUILD"
)
}
\arguments{
\item{pkg_root}{Character string.
Package root directory.}

\item{clean}{Logical.
Clean output that were automatically generated
by previous builds
before initiating the current build.
See \code{\link[=pkg_clean]{pkg_clean()}} for details.}

\item{minimal}{Logical.
If \code{TRUE}, minimal check, build, and install
using \verb{R CMD} will be performed,
ignoring all the other function arguments.
If \code{FALSE}, a more thorough check is performed
with additional options.
See \code{\link[=pkg_build_minimal]{pkg_build_minimal()}} for details.}

\item{style}{Logical.
Style \code{R} scripts and \code{R} Markdown files.
See \code{\link[=pkg_style]{pkg_style()}} for details.}

\item{data}{Logical.
Generate data from \code{data_raw}.
See \code{\link[=pkg_data_raw]{pkg_data_raw()}} for details.}

\item{render}{Logical.
Render \code{R} scripts and \code{R} Markdown files.
See \code{\link[=pkg_render]{pkg_render()}} for details.}

\item{readme}{Logical.
Render \code{README.Rmd}.
Ignored if \code{render = FALSE}.}

\item{vignettes}{Logical.
Render \verb{vignettes/*.Rmd}.
Ignored if \code{render = FALSE}.}

\item{tests}{Logical.
Render \verb{tests/testhat/*.R}.
Ignored if \code{render = FALSE}.}

\item{tests2vignettes}{Logical.
Copy and spin \verb{tests/testhat/*.R}
to \verb{vignettes/tests/*.Rmd}.
It is assumed that \verb{tests/testhat/*.R}
are written using \code{Roxygen} comments.
See \code{\link[=pkg_tests2vignettes]{pkg_tests2vignettes()}} for details.}

\item{pkgdown}{Logical.
Build \code{pkgdown} site.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.}

\item{ncores}{Integer.
Number of cores to use if \code{par = TRUE}.
If unspecified, defaults to \code{detectCores() - 1}.}

\item{git}{Logical.
Set up a git repository.}

\item{github}{Logical.
Set up and push to a github repository.}

\item{commit_msg}{Character string.
Git commit message.}
}
\description{
Builds the \code{R} package on \code{pkg_root} by:
\itemize{
\item styling the \code{R} scripts and \code{R} Markdown files,
\item building package data by running \code{R} scripts stored in \code{data_raw},
\item rendering the \code{R} scripts and \code{R} Markdown files using \code{\link[rmarkdown:render]{rmarkdown::render()}}, and
\item building \code{pkgdown} site.
}
}
\examples{
\dontrun{
pkg_build(
  pkg_root = "~/boilerplatePackage",
  minimal = TRUE
)
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
