% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_build.R
\name{pkg_build}
\alias{pkg_build}
\title{Build Package.}
\usage{
pkg_build(
  pkg_root = getwd(),
  style = FALSE,
  data = FALSE,
  render = FALSE,
  readme = FALSE,
  vignettes = FALSE,
  tests = FALSE,
  pkgdown = FALSE,
  par = TRUE,
  ncores = NULL
)
}
\arguments{
\item{pkg_root}{Character string.
Package root directory.}

\item{style}{Logical.
Style R scripts and R Markdown files.}

\item{data}{Logical.
Generate data from \code{data_raw}.}

\item{render}{Logical.
Render \code{R} scripts and \code{R} Markdown files.}

\item{readme}{Logical.
Render \code{README.Rmd}.
Ignored if \code{render = FALSE}.}

\item{vignettes}{Logical.
Render \verb{vignettes/*.Rmd}.
Ignored if \code{render = FALSE}.}

\item{tests}{Logical.
Render \verb{tests/testhat/*.R}.
Ignored if \code{render = FALSE}.}

\item{pkgdown}{Logical.
Build \code{pkgdown} site.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.}

\item{ncores}{Integer.
Number of cores to use if \code{par = TRUE}.
If unspecified, defaults to \code{detectCores() - 1}.}
}
\description{
Builds the \code{R} package on \code{pkg_root} by:
\itemize{
\item styling the \code{R} scripts and \code{R} Markdown files,
\item building package data by running \code{R} scripts stored in \code{data_raw},
\item rendering the \code{R} scripts and \code{R} Markdown files using \code{\link[rmarkdown:render]{rmarkdown::render()}}, and
\item building \code{pkgdown} site.
}
}
\details{
Some features of this function requires
}
\examples{
\dontrun{
pkg_build(
  pkg_root = getwd(),
  style = FALSE,
  data = FALSE,
  render = FALSE,
  readme = FALSE,
  vignettes = FALSE,
  tests = FALSE,
  pkgdown = FALSE,
  par = FALSE
)
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
