% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_data.R
\name{pkg_data}
\alias{pkg_data}
\title{Create a Boilerplate Package Data File.}
\usage{
pkg_data(pkg_root, msg = "Data script file path:")
}
\arguments{
\item{pkg_root}{Character string.
Package root directory.}

\item{msg}{Character string.
Optional message.
If supplied,
prints \verb{msg dir/fn}.}
}
\description{
Creates a boilerplate package data file.
}
\details{
Note that if \code{\link[jeksterslabRpkg:pkg_create]{jeksterslabRpkg::pkg_create()}} is used,
this function will be called.
}
\examples{
\dontrun{
pkg_data(
  pkg_root = "boilerplatePackage"
)
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
