% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_data.R
\name{pkg_data}
\alias{pkg_data}
\title{Create Boilerplate Package Data File.}
\usage{
pkg_data(pkg_dir = getwd(), pkg_name)
}
\arguments{
\item{pkg_dir}{Directory where the package is initialized.}

\item{pkg_name}{Character string.
Package name.}
}
\description{
Create Boilerplate Package Data File.
}
\examples{
pkg_data(pkg_dir = tempdir(), pkg_name = "boilerplatePackage")
}
\author{
Ivan Jacob Agaloos Pesigan
}
