% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_description.R
\name{pkg_description}
\alias{pkg_description}
\title{Create Package DESCRIPTION.}
\usage{
pkg_description(pkg_dir = getwd(), pkg_name, input_file = NULL, add = NULL)
}
\arguments{
\item{pkg_dir}{Directory where the package is initialized.}

\item{pkg_name}{Character string.
Package name.}

\item{input_file}{csv file containing \code{DESCRIPTION} fields and entries.}

\item{add}{Character string.
Additional entries to the \code{DESCRIPTION} file
not included in \code{input_file}}
}
\description{
Creates package \code{DESCRIPTION}
by extracting information
from an external \code{csv} \code{input_file}.
See \code{system.file("extdata", "DESCRIPTION.csv", package = "jeksterslabRpkg", mustWork = TRUE)}
For the \code{input_file} template.
THIS FUNCTION OVERWRITES AN EXISTING \code{DESCRIPTION} FILE
IN THE SPECIFIED PACKAGE ROOT DIRECTORY.
USE WITH CAUTION.
}
\examples{
pkg_description(pkg_dir = tempdir(), pkg_name = "boilerplatePackage")
}
\author{
Ivan Jacob Agaloos Pesigan
}
