% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_gitignore.R
\name{pkg_gitignore}
\alias{pkg_gitignore}
\title{Create Boilerplate Package .gitignore.}
\usage{
pkg_gitignore(pkg_dir = getwd(), pkg_name, add = NULL)
}
\arguments{
\item{pkg_dir}{Directory where the package is initialized.}

\item{pkg_name}{Character string.
Package name.}

\item{add}{Character string.
Entries to the \code{.gitignore}
in addition to the boilerplate example.}
}
\description{
Create Boilerplate Package .gitignore.
}
\examples{
pkg_gitignore(pkg_dir = tempdir(), pkg_name = "boilerplatePackage")
}
\author{
Ivan Jacob Agaloos Pesigan
}
