% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_namespace.R
\name{pkg_namespace}
\alias{pkg_namespace}
\title{Create Boilerplate Package NAMESPACE.}
\usage{
pkg_namespace(pkg_dir = getwd(), pkg_name, add = NULL)
}
\arguments{
\item{pkg_dir}{Directory where the package is initialized.}

\item{pkg_name}{Character string.
Package name.}

\item{add}{Character string.
Entries to the \code{NAMESPACE}
in addition to the boilerplate example.}
}
\description{
Create Boilerplate Package NAMESPACE.
}
\examples{
pkg_namespace(pkg_dir = tempdir(), pkg_name = "boilerplatePackage")
}
\author{
Ivan Jacob Agaloos Pesigan
}
