% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_namespace.R
\name{pkg_namespace}
\alias{pkg_namespace}
\title{Create a Boilerplate Package \code{NAMESPACE}.}
\usage{
pkg_namespace(pkg_root, add = NULL, msg = "NAMESPACE file path:")
}
\arguments{
\item{pkg_root}{Character string.
Package root directory.}

\item{add}{Character string.
Entries to the \code{NAMESPACE}
in addition to the boilerplate example.}

\item{msg}{Character string.
Optional message.
If supplied,
prints \verb{msg dir/fn}.}
}
\description{
Creates a boilerplate package \code{NAMESPACE}.
}
\details{
Note that if \code{\link[jeksterslabRpkg:pkg_create]{jeksterslabRpkg::pkg_create()}} is used,
this function will be called.

If you are going to document, check, and build your package using \code{devtools},
you may delete the boilerplate \code{NAMESPACE} file.

\strong{THIS FUNCTION OVERWRITES AN EXISTING \code{NAMESPACE} FILE
IN THE SPECIFIED PACKAGE ROOT DIRECTORY.
USE WITH CAUTION.}
}
\examples{
\dontrun{
pkg_namespace(
  pkg_root = "~/boilerplatePackage"
)
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
