% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_news.R
\name{pkg_news}
\alias{pkg_news}
\title{Create a Boilerplate Package NEWS.md File.}
\usage{
pkg_news(pkg_root, input_file = NULL, msg = "NEWS.md file path:")
}
\arguments{
\item{pkg_root}{Character string.
Package root directory.}

\item{input_file}{Character string.
\code{YAML} file containing \code{DESCRIPTION} fields and entries.}

\item{msg}{Character string.
Prints optional message and output directory.}
}
\description{
Creates a boilerplate package NEWS.md file.
}
\details{
Note that if \code{\link[=pkg_create]{pkg_create()}} is used,
this function will be called.
}
\examples{
\dontrun{
pkg_news(
  pkg_root = "~/boilerplatePackage",
  input_file = "DESCRIPTION.yml"
)
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
