% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_pkgdown.R
\name{pkg_pkgdown}
\alias{pkg_pkgdown}
\title{Create a Boilerplate Package \verb{_pkgdown.yml}.}
\usage{
pkg_pkgdown(pkg_root, input_file = NULL, msg = "_pkgdown.yml file path:")
}
\arguments{
\item{pkg_root}{Character string.
Package root directory.}

\item{input_file}{Character string.
\code{YAML} file containing \code{DESCRIPTION} fields and entries.}

\item{msg}{Character string.
Prints optional message and output directory.}
}
\description{
Creates a \verb{_pkgdown.yml} build file.
}
\details{
Note that if \code{\link[=pkg_create]{pkg_create()}} is used,
and the argument \code{pkgdown = TRUE} is specified,
this function will be called.

\strong{THIS FUNCTION OVERWRITES AN EXISTING \verb{_pkgdown.yml} FILE
IN THE SPECIFIED PACKAGE ROOT DIRECTORY.
USE WITH CAUTION.}
}
\examples{
\dontrun{
pkg_pkgdown(
  pkg_root = "~/boilerplatePackage",
  input_file = "DESCRIPTION.yml"
)
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
