% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_r.R
\name{pkg_r}
\alias{pkg_r}
\title{Create a Boilerplate Package \code{R} Script.}
\usage{
pkg_r(pkg_root, msg = "z.R file path:")
}
\arguments{
\item{pkg_root}{Character string.
Package root directory.}

\item{msg}{Character string.
Optional message.
If supplied,
prints \verb{msg dir/fn}.}
}
\description{
Creates a boilerplate package \code{R} script.
}
\details{
Note that if \code{\link[jeksterslabRpkg:pkg_create]{jeksterslabRpkg::pkg_create()}} is used,
this function will be called.
}
\examples{
\dontrun{
pkg_r(
  pkg_root = getwd()
)
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
