% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_r.R
\name{pkg_r}
\alias{pkg_r}
\title{Create Boilerplate Package R Script.}
\usage{
pkg_r(pkg_dir = getwd(), pkg_name)
}
\arguments{
\item{pkg_dir}{Directory where the package is initialized.}

\item{pkg_name}{Character string.
Package name.}
}
\description{
Create Boilerplate Package R Script.
}
\examples{
pkg_r(pkg_dir = tempdir(), pkg_name = "boilerplatePackage")
}
\author{
Ivan Jacob Agaloos Pesigan
}
