% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_rbuildignore.R
\name{pkg_rbuildignore}
\alias{pkg_rbuildignore}
\title{Create Boilerplate Package .Rbuildignore.}
\usage{
pkg_rbuildignore(pkg_dir = getwd(), pkg_name, add = NULL)
}
\arguments{
\item{pkg_dir}{Directory where the package is initialized.}

\item{pkg_name}{Character string.
Package name.}

\item{add}{Character string.
Entries to the \code{.Rbuildignore}
in addition to the boilerplate example.}
}
\description{
Create Boilerplate Package .Rbuildignore.
}
\author{
Ivan Jacob Agaloos Pesigan
}
