% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_rbuildignore.R
\name{pkg_rbuildignore}
\alias{pkg_rbuildignore}
\title{Create a Boilerplate Package \code{.Rbuildignore} File.}
\usage{
pkg_rbuildignore(pkg_root, add = NULL, msg = ".Rbuildignore file path:")
}
\arguments{
\item{pkg_root}{Character string.
Package root directory.}

\item{add}{Character string.
Entries to the \code{.Rbuildignore}
in addition to the boilerplate example.}

\item{msg}{Character string.
Optional message.
If supplied,
prints \verb{msg dir/fn}.}
}
\description{
Creates a boilerplate package \code{.Rbuildignore} file.
}
\details{
Note that if \code{\link[=pkg_create]{pkg_create()}} is used,
this function will be called.

If \code{.Rbuildignore} exists in the specified package root directory,
it will \strong{NOT} be overwritten.
Additional entries in the \code{add} argument will be appended.
}
\examples{
\dontrun{
pkg_rbuildignore(
  pkg_root = "~/boilerplatePackage"
)
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
