% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_rd.R
\name{pkg_rd}
\alias{pkg_rd}
\title{Create a Boilerplate Package \code{Rd} Function Documentation.}
\usage{
pkg_rd(pkg_dir = getwd(), pkg_name)
}
\arguments{
\item{pkg_dir}{Character string.
Directory where the package is initialized.}

\item{pkg_name}{Character string.
Package name.}
}
\description{
Creates a boilerplate package \code{Rd} function documentation.
}
\details{
Note that if \code{\link[jeksterslabRpkg:pkg_create]{jeksterslabRpkg::pkg_create()}} is used,
this function will be called.

If you are going to document, check, and build your package using \code{devtools},
you may delete the \code{Rd} function documentation file generated by this function.
}
\examples{
\dontrun{
pkg_rd(
  pkg_dir = getwd(),
  pkg_name = "boilerplatePackage"
)
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
