% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_rd.R
\name{pkg_rd}
\alias{pkg_rd}
\title{Create Boilerplate Package RD Function Documentation.}
\usage{
pkg_rd(pkg_dir = getwd(), pkg_name)
}
\arguments{
\item{pkg_dir}{Directory where the package is initialized.}

\item{pkg_name}{Character string.
Package name.}
}
\description{
Create Boilerplate Package RD Function Documentation.
}
\examples{
pkg_rd(pkg_dir = tempdir(), pkg_name = "boilerplatePackage")
}
\author{
Ivan Jacob Agaloos Pesigan
}
