% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_readme.R
\name{pkg_readme}
\alias{pkg_readme}
\title{Create Package README.}
\usage{
pkg_readme(pkg_dir = getwd(), pkg_name, input_file = NULL)
}
\arguments{
\item{pkg_dir}{Directory where the package is initialized.}

\item{pkg_name}{Character string.
Package name.}

\item{input_file}{csv file containing \code{DESCRIPTION} fields and entries.}
}
\description{
Creates package \code{README.Rmd}
by extracting information
from an external \code{csv} \code{input_file}.
See \code{system.file("extdata", "DESCRIPTION.csv", package = "jeksterslabRpkg", mustWork = TRUE)}
For the \code{input_file} template.
THIS FUNCTION OVERWRITES AN EXISTING \code{README.Rmd} FILE
IN THE SPECIFIED PACKAGE ROOT DIRECTORY.
USE WITH CAUTION.
}
\examples{
pkg_readme(pkg_dir = tempdir(), pkg_name = "boilerplatePackage")
}
\author{
Ivan Jacob Agaloos Pesigan
}
