% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_rproj.R
\name{pkg_rproj}
\alias{pkg_rproj}
\title{Create Package .Rproj.}
\usage{
pkg_rproj(pkg_dir = getwd(), pkg_name)
}
\arguments{
\item{pkg_dir}{Directory where the package is initialized.}

\item{pkg_name}{Character string.
Package name.}
}
\description{
Create Package .Rproj.
}
\examples{
pkg_rproj(pkg_dir = tempdir(), pkg_name = "boilerplatePackage")
}
\author{
Ivan Jacob Agaloos Pesigan
}
