% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_rproj.R
\name{pkg_rproj}
\alias{pkg_rproj}
\title{Create a Package \code{.Rproj} File.}
\usage{
pkg_rproj(pkg_root = getwd(), msg = "Rproj file path:")
}
\arguments{
\item{pkg_root}{Character string.
Package root directory.}

\item{msg}{Character string.
Optional message.
If supplied,
prints \verb{msg dir/fn}.}
}
\description{
Creates a package \code{.Rproj} file.
}
\details{
Note that if \code{\link[=pkg_create]{pkg_create()}} is used,
this function will be called.

\strong{THIS FUNCTION OVERWRITES AN EXISTING \code{.Rproj} FILE
IN THE SPECIFIED PACKAGE ROOT DIRECTORY.
USE WITH CAUTION.}
}
\examples{
\dontrun{
pkg_rproj(
  pkg_root = "~/boilerplatePackage"
)
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
