% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_style.R
\name{pkg_style}
\alias{pkg_style}
\title{Style \code{R} and \code{R} Markdown Files in the Package Directory.}
\usage{
pkg_style(pkg_root = getwd(), par = TRUE, ncores = NULL)
}
\arguments{
\item{pkg_root}{Character string.
Package root directory.}
}
\description{
Styles \code{R} and \code{R} Markdown files in the package directory.
}
\examples{
\dontrun{
pkg_style(
  pkg_root = "~/boilerplatePackage",
  par = FALSE
)
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
