% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_style.R
\name{pkg_style}
\alias{pkg_style}
\title{Style \code{R} and \code{R} Markdown Files in the Package Directory}
\usage{
pkg_style(pkg_root = getwd(), par = FALSE, ncores = NULL)
}
\arguments{
\item{pkg_root}{Character string.
Package root directory.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.}

\item{ncores}{Integer.
Number of cores to use if \code{par = TRUE}.
If unspecified, defaults to \code{detectCores() - 1}.}
}
\description{
Styles \code{R} and \code{R} Markdown files in the package directory.
}
\examples{
\dontrun{
pkg_style(
  pkg_root = "~/boilerplatePackage",
  par = FALSE
)
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
