% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_test.R
\name{pkg_test}
\alias{pkg_test}
\title{Create Boilerplate Package Test File.}
\usage{
pkg_test(pkg_dir = getwd(), pkg_name)
}
\arguments{
\item{pkg_dir}{Directory where the package is initialized.}

\item{pkg_name}{Character string.
Package name.}
}
\description{
Create Boilerplate Package Test File.
}
\author{
Ivan Jacob Agaloos Pesigan
}
