% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_test.R
\name{pkg_test}
\alias{pkg_test}
\title{Create a Boilerplate Package Test File.}
\usage{
pkg_test(pkg_root, msg = "test_z.R file path:")
}
\arguments{
\item{pkg_root}{Character string.
Package root directory.}

\item{msg}{Character string.
Optional message.
If supplied,
prints \verb{msg dir/fn}.}
}
\description{
Creates a boilerplate package test file.
}
\details{
Note that if \code{\link[jeksterslabRpkg:pkg_create]{jeksterslabRpkg::pkg_create()}} is used,
this function will be called.
}
\examples{
\dontrun{
pkg_test(
  pkg_root = getwd()
)
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
