% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_travis.R
\name{pkg_travis}
\alias{pkg_travis}
\title{Create Boilerplate Package .travis.yml.}
\usage{
pkg_travis(pkg_dir = getwd(), pkg_name, add = NULL)
}
\arguments{
\item{pkg_dir}{Directory where the package is initialized.}

\item{pkg_name}{Character string.
Package name.}

\item{add}{Character string.
Entries to the \code{.gitignore}
in addition to the boilerplate example.}
}
\description{
Creates .travis.yml with \code{codecov}.
}
\author{
Ivan Jacob Agaloos Pesigan
}
