% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_vignette.R
\name{pkg_vignette}
\alias{pkg_vignette}
\title{Create a Boilerplate Package Vignette File}
\usage{
pkg_vignette(pkg_root, msg = "z.Rmd file path:")
}
\arguments{
\item{pkg_root}{Character string.
Package root directory.}

\item{msg}{Character string.
Optional message.
If supplied,
prints \verb{msg dir/fn}.}
}
\description{
Creates a boilerplate package vignette file.
}
\details{
Note that if \code{\link[=pkg_create]{pkg_create()}} is used,
this function will be called.
}
\examples{
\dontrun{
pkg_vignette(
  pkg_root = "~/boilerplatePackage"
)
}
}
\author{
Ivan Jacob Agaloos Pesigan
}
