% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_aliases.R
\name{rpkg}
\alias{rpkg}
\title{Alias for styler::style_pkg(); devtools::check(); devtools::install()}
\usage{
rpkg()
}
\description{
Alias for
\code{\link[styler:style_pkg]{styler::style_pkg()}},
\code{\link[devtools:check]{devtools::check()}},
\code{\link[devtools:install]{devtools::install()}}
with default arguments.
}
\seealso{
\code{\link[styler:style_pkg]{styler::style_pkg()}}

\code{\link[devtools:check]{devtools::check()}}

\code{\link[devtools:install]{devtools::install()}}
}
