% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_aliases.R
\name{rrender}
\alias{rrender}
\title{Alias for jeksterslabRutils::util_render()}
\usage{
rrender(
  dir = getwd(),
  recursive = TRUE,
  files = NULL,
  par = TRUE,
  ncores = NULL
)
}
\description{
Alias for \code{\link[jeksterslabRutils:util_render]{jeksterslabRutils::util_render()}}
with the following arguments:
\itemize{
\item \code{dir = getwd()}
\item \code{recursive = TRUE}
\item \code{files = NULL}
\item \code{par = TRUE}
\item \code{ncores = NULL}
}
}
\seealso{
\code{\link[jeksterslabRutils:util_render]{jeksterslabRutils::util_render()}}
}
