% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_aliases.R
\name{rstyle}
\alias{rstyle}
\title{Alias for jeksterslabRutils::util_style()}
\usage{
rstyle(dir = getwd(), recursive = TRUE, par = TRUE, ncores = NULL)
}
\description{
Alias for \code{\link[jeksterslabRutils:util_style]{jeksterslabRutils::util_style()}}
with the following arguments:
\itemize{
\item \code{dir = getwd()}
\item \code{recursive = TRUE}
\item \code{par = TRUE}
\item \code{ncores = NULL}
}
}
\seealso{
\code{\link[jeksterslabRutils:util_style]{jeksterslabRutils::util_style()}}
}
