% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_aliases.R
\name{rstyle}
\alias{rstyle}
\title{Aliases jeksterslabRpkg::util_style() Recursive and Parallel}
\usage{
rstyle(dir = getwd(), recursive = TRUE, par = TRUE, ncores = NULL)
}
\arguments{
\item{dir}{Character string.
Directory.}

\item{recursive}{Logical.
If \code{TRUE},
recursively style all \code{R} scripts (\code{.R}, \code{.r}) and
R Markdown files (\code{.Rmd}, \code{.rmd})
in \code{dir}.}

\item{par}{Logical.
If \code{TRUE}, use multiple cores.}

\item{ncores}{Integer.
Number of cores to use if \code{par = TRUE}.
If unspecified, defaults to \code{detectCores() - 1}.}
}
\description{
\code{jeksterslabRpkg::util_style( dir = getwd(), recursive = TRUE, par = TRUE, ncores = NULL )}
}
